requirements.txt
absl-py==2.1.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
dill==0.3.8
filelock==3.15.4
Flask==3.0.3
fonttools==4.53.1
fsspec==2024.6.1
h5py==3.12.1
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
keras==2.11.0
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
ml_dtypes==0.5.0
mpmath==1.3.0
namex==0.0.8
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
opencv-python==4.10.0.84
optree==0.13.0
packaging==24.1
pandas==2.2.2
pillow==10.4.0
psutil==6.0.0
py-cpuinfo==9.0.0
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
rich==13.9.2
scipy==1.14.1
seaborn==0.13.2
six==1.16.0
sympy==1.13.2
torch==2.4.0
torchvision==0.19.0
tqdm==4.66.5
triton==3.0.0
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.2.86
ultralytics-thop==2.0.6
urllib3==2.2.2
Werkzeug==3.0.4

tensorflow~=2.17.0
keras-nightly~=3.6.0.dev2024100503
Flask-Cors~=5.0.0